Check if at least a driver is supported/enabled.
I had noticed that it is possible that FluidSynth does not compile if none of the audio out drivers is supported. I would like to suggest introduce a check, in the same manner it has been done for the MIDI input drivers. Actually, the absence of an audio driver does not mean that FluidSynth is not usable, it will always act as a rendering engine while the user will generate the samples with fluid_synth_write_float()/fluid_synth_write_s16() and play them with the audio API provided by his operating system.