go back to pkg-config-less assetloader build.
It has to be done so because arm64-v8a build actually does not "complete"
due to linking failure at `fluidsynth` executable (with libiconv, which
is now supposed to be part of AOSP system library and therefore harmless)
and thus prevents "make install".

To make it build, now this build avoids dependency on "fluidsynth.h" and
include individual headers directly.