Add OpenSLES and Oboe audio drivers for Android.
This is a simplified changeset from https://github.com/FluidSynth/fluidsynth/pull/464 where all the detailed discussion and individual changes are tracked.

`android` directory contains a lot of hints and scripts for building
fluidsynth for Android, as well as `README.Android.md`.

There are couple of fixes to existing sources for clang compilation.