Remove (incomplete) support for `new_fluid_audio_driver2()`.
I had some misunderstandings on how `new_fluid_audio_driver2()` are
(not) compatible with `new_fluid_audio_driver()` - it was not just
about passing `fluid_synth_process` but also about how to handle
different buffers. No wonder why it didn't work well with callbacks
(it was disabled in opensles, and seemed to have never worked fine
on oboe either). There are handful of adrivers without it anyways.

Remove any relevant code for this custom callback for those drivers.
(Someone or /me might be interested in supporting it.)