fluid_opensles.c : lots of simplification by removing non-callback mode.
I had some confusion. I implemented everything in the principle that
all those "callbacks" are optional i.e. it may or may not be provided,
for each. And by those "callbacks", I mean, are (1) fluidsynth custom
callbacks to fill the buffers, and (2) OpenSLES callbacks that apps
(in this case, my adriver) registers to OpenSLES. I also made use of
it to identify what was going wrong with the audio output (when it was).

The latter is totally optional, so I could safely remove it, and it
significantly simplifies the code (in this change).