Workaround for jack sample rate mismatch
During creation of a jack audio driver, it is checked whether the
sample-rate of the settings object matches jack's rate. If not, it was
adjusted previously via fluid_synth_set_sample_rate(). Due to the
deprecation of that function and removal of real-time capability of
synth.sample-rate setting, a regression was introduced in
5fbddcecc39476e471308c737ef8124ae3c73eeb causing the synth's sample-rate
to be not updated.
This workaround obtains the synth via the settings instance and for now
calls the deprecated sample-rate set function.