fluid_jack: prevent using a different setting at synth and jack
When Jack plugin is set to use multiple stereo outputs, if
set commands (even at the config file) is changing the number
of audio and/or effect effect channels/groups, it may happen
that the settings used by fluid_synth to be different than the
ones used by fluid_jack driver.

This cause errors at the fluid_synth_process() function,
making it return FLUID_FAILED, causing the Jack thread
to stop.

When this happens, the entire jack plugin stops working, which
prevents connecting the fluidsynth outputs. So no audio
goes to the outputs.

The real fix would be to add a runtime update that would be
changing the fluid_synth settings when this happens. However,
a change like that is not trivial and might require some
changes at the API.

So, instead, check if the settings will cause problems
in advance, not leting the plugin initialize on such cases
and printing an error message.