Fix incorrect maximum excursion of modulators (#1152)
Previously, fluidsynth used an incorrect range for modulators. I.e. the maximum excursion was `1.0`. The soundfont spec however specifies the maximum excursion to be `127/128`.

By implementing this change the special treatment of MIDI RP-036 becomes redundant because with this new range of `128` the center of 63.5 can be correctly represented, i.e. `(127/128) * 64 = 63.5` or normalized as fluidsynth does `64/128 = 0.5`.

Now, only the pitch modulator remains which uses a custom range of `0x4000`, as this is the only modulator we implement with true 14-bit resolution.

I did not add any more unit tests for this because primarily I'm lacking time. And also because it looks fairly trivial to me. Yet, I think this change should go into 2.3.0.