Fix issue with CMake Xcode generator
There is a known issue with CMake and its Xcode generator when
the target has no sources of its own -- see

https://gitlab.kitware.com/cmake/cmake/-/issues/23688

A workaround is to define an empty source file that can be used
to trigger the build path for the framework. Without this hack,
the framework is missing the shared library component.

With this change, the following now runs successfully:

``
% cmake -G Xcode -Denable-sdl2=NO -Denable-readline=NO -Denable-dbus=NO -DCMAKE_INSTALL_PREFIX=$PWD ..
% xcodebuild
```