Only run one FluidSynth service at a time
Trying to run a second server will fail unless the user has configured
a non-default port.  This happens every time a user switches user,
because the display manager is implemented as a systemd user session.
Use a lock file to avoid that happening.

This gets rid of an error message for most users, but forces advanced users
to override the systemd configuration.  Such users should be smart enough to
see the condition failure in the logs, check fluidsynth.service, and understand
the hint about `systemctl edit`.