Make *_SUPPORT macros consistently be 1 or undefined.
The various <MODULE>_FOUND variables in CMake evaluate to TRUE/FALSE.
When these are patched through to preprocessor defines in config.cmake,
they become either TRUE or remain undefined. This happens to work with
the various #if <MODULE>_SUPPORT checks in the code only because GLib
defines TRUE to 1. This causes hard to track down missing definitions
when not including GLib headers.

Adjust all places to check for <MODULE>_FOUND and then set the
<MODULE>_SUPPORT variable to 1.

Also update the example in README.cmake.md and clean up some missing
conditions in else () and endif ().