Add an "embedded" OS abstraction that contains mostly stubs.
This abstraction is intended to be used when pure embedding in another
project is desired. It does not support any threading, has no atomics
or time functions and does not support LADSPA or InstPatch. It is
therefore only useful for standalone rendering of audio, which is often
the only thing needed when embedding into another project.

Audio drivers are in principle usable as long as they don't need a
separate audio thread, like PipeWire.