Fix missing API Reference with Doxygen >= 1.9.8.
Due to a name conflict with C++ modules, the term "modules" was replaced
with "topics" in Doxygen 1.9.8. As "modules" is used for the tab type
in the customized layout, that tab is then not produced at all anymore.

Process the custom layout file to set the tab type to either "modules"
or "topics" depending on the Doxygen version to restore the tab.