Set the Doxyfile config HTML_COLORSTYLE to LIGHT.
This option is present in versions of Doxygen >= 1.9.5 and defaults to
AUTO_LIGHT, meaning it will default to a light style and switch to a
dark one in browsers configured to prefer dark mode.

The dark mode completely breaks with the custom stylesheet as that one
overrides various, but not all, colors with ones from a light style.

Force the use of the light style without switching to dark via explicit
LIGHT config until the custom theme can be adapted.