Add an initial implementation of a C++11 based OSAL.
It uses locking primitives and threading from C++11.

Atomics are implemented with a global mutex lock. This is permitted by
C++11 and fulfills the semantics, but is obviously suboptimal. Making
use of std::atomic in the code base will require more work.

Some functions are still stubs here, namely the file access related ones
and the command line parsing.