Add C99 conformant [v]snprintf() wrappers when not using GLib.
When not using GLib, use the non-conformant _vsnprintf() that does not
properly terminate on overflow, but wrap it in an inline that ensures
termination is always done. Then implement sprintf() on top of the now
conformant vsnprintf() wrapper or an already conformant system version.

There already is a unit test that validates proper behaviour.

This seems to be the last GLib dependency in fluidsynth_priv.h.