Eliminate use of fluid_file_test for config file sourcing.
This removes the use of the unexported private function fluid_file_test
from the main binary. The pattern of testing for file existence before
opening is generally prone to time-of-check-to-time-of-use races and
GLib explicitly advises against it.

Instead, try to source the possible configuration files directly until
one works or all paths have been tried.

Note that this produces a change in behaviour when configuration files
are present but inaccessible (due to lack of permission or broken
symlinks for example). Previously it was enough for a higher priority
file to be present, even if it later failed to be sourced, to block
further iteration. Now, inaccessible files will just be skipped.