Fix exclusion of fluid_thread_self_set_prio in embedded OSAL. (#1587)
This is a rework of the change in #1575. The embedded OSAL may be used
on non-POSIX platforms as well, so the exclusion of the
fluid_thread_self_set_prio definition needs to apply to all of them.

Also the declaration of the function doesn't need to be excluded as it
is defined in all cases, just possibly as a stub.