Add limiter support (with limiter from signalsmith-audio-basics)
This adds limiter support by using Signalsmith Audio Basics C++
header-only library (see https://signalsmith-audio.co.uk/code/ and
https://github.com/Signalsmith-Audio/basics)

Path to Signalsmith basics lib have to be supplied to cmake in the
environment variable SIGNALSMITH_AUDIO_BASICS_PATH

The limiter is applied in fluid_rvoice_mixer_process_fx after
chorus and reverb effects. In the audio drivers that mix effects
and dry audio in one buffer, the limiting will be applied to the
resulting buffer. In audio drivers that do this separately (for
example, Jack driver), the limiting will be applied only to the
dry buffer